feat: support speaker and mic for Airbox Orin
On the Airbox Orin, the speaker and mic need the audio codec to work.
The audio codec voltage input is controlled by PAC.06(GPIO09). The
speaker also needs the audio amplifier to work. The amplifier enable pin
is controlled by PH.00(GPIO13).

In this change, PAC.06 and PH.00 are configured as outputs in pinmux,
and the dt configs of the audio codec and the audio amplifier are
attached as a patch file.

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>