fix: use orin nano devkit super dtb
By default, when using the Nvidia Jetson Linux SD card image[0] with
this firmware, the device tree blob provided by this firmware was loaded
when the system starts up, which meant the patched[1] version of
`tegra234-p3768-0000+p3767-0005-nv.dtb` was loaded. To the device tree,
the value of `model` was "Radxa Orin C200" and that of `compatible` was
"nvidia,p3768-0000+p3767-0005 nvidia,p3767-0005 nvidia,tegra234".

When users want to use cameras on the Nvidia Jetson Linux, they need to
execute `sudo /opt/nvidia/jetson-io/jetson-io.py` to enter a menu to
configure the CSI Connector[2]. Before this change, the command would
exit immediately without any error message. When running `sudo
/opt/nvidia/jetson-io/confi-by-pin.py`, there will be a message saying
"No DTB found for Radxa Orin C200". When inspecting the source code of
`/opt/nvidia/jetson-io/Jetson/board.py`, we can see that in
`/opt/nvidia/jetson-io/Utils/dtc.py`:

```
...
def find_compatible_dtb_files(compat, model, path):
    dtbs = []
    for dtb in glob.glob(os.path.join(path, '*.dtb')):
        if compat != get_compatible(dtb):
            continue
        if model != get_model(dtb):
            continue
        dtbs.append(dtb)
    if not dtbs:
        return None
    return dtbs
...
```

The `model` and `compatible` of the device tree, given by the firmware,
are used to find a dtb file in `/boot/dtb/`. In the official SD card
image. There is only one file in `/boot/dtb`, which is
`/boot/dtb/kernel_tegra234-p3768-0000+p3767-0005-nv-super.dtb`. After
decompiling the blob, we can see that the value of `model` and
`compatible` of that blob are "NVIDIA Jetson Orin Nano Engineering
Reference Developer Kit Super" and
"nvidia,p3768-0000+p3767-0005-super nvidia,p3767-0005 nvidia,tegra234".
It means that the value of `model` and `compatible` have to be the same
as those of `kernel_tegra234-p3768-0000+p3767-0005-nv-super.dtb`,
otherwise the jetson-io utilities cannot be used.

This commit modifies the Makefile so the device tree of Orin Nano Super
is used, and removes a patch overwriting the model name.

[0]:
https://developer.nvidia.com/downloads/embedded/L4T/r36_Release_v4.4/jp62-r1-orin-nano-sd-card-image.zip
[1]:
https://github.com/radxa/c200-bootupd/blob/bee301f02a23682b6f48ab4893e447f5b878f8ac/patches/t23x-public-dts/c200/0001-t23x-dts-update-product-model.patch
[2]:
https://docs.nvidia.com/jetson/archives/r36.4.4/DeveloperGuide/HR/ConfiguringTheJetsonExpansionHeaders.html#configuring-the-csi-connector

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>