fix: allows compile without any patch
When the directory `patches/t23x-public-dts/$(PRODUCT)` is empty, it
will be removed automatically by git. Then, due to the lack of this
directory, the `git am` command will fail, so `make build` will fail.
This commit adds a conditional statement checking if the directory
`patches/t23x-public-dts/$(PRODUCT)` exists, and executes the `git am`
command only if the directory exists.

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>