feat: improve Makefile
Requiring setting the variables before building can avoid the situation
that, in the first time building, the default configurations are not what
users want.

Some paths are hardcoded with c200/*. Now this repo supports both Airbox
Orin and C200, we should use the PRODUCT variable to replace the
hardcoded part of the paths.

For consistency, README is updated to introduce the possible values of
PRODUCT and BOARDSKU.

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>